+++ mycodo/inputs/grove_temp_humidity.py
@@ -6,11 +6,9 @@ from flask_babel import lazy_gettext
 import sys
 sys.path.insert(1, '/usr/local/bin')
 
-from mycodo.databases.models import Output
 from mycodo.inputs.base_input import AbstractInput
 from mycodo.inputs.sensorutils import calculate_dewpoint
 from mycodo.inputs.sensorutils import calculate_vapor_pressure_deficit
-from mycodo.utils.database import db_retrieve_table_daemon
 
 # Measurements
 measurements_dict = {
@@ -35,14 +33,16 @@ measurements_dict = {
 # Input information
 INPUT_INFORMATION = {
     'input_name_unique': 'GROVE_TEMP_AND_HUMIDITY',
-    'input_manufacturer': 'Grove',
-    'input_name': 'DHT11_DHT22',
+    'input_manufacturer': 'Grove Pi',
+    'input_name': 'DHT11/DHT22',
     'input_library': 'grovepi',
     'measurements_name': 'Humidity/Temperature',
     'measurements_dict': measurements_dict,
-    'url_datasheet': 'https://wiki.seeedstudio.com/Grove-TemperatureAndHumidity_Sensor/',
-    'url_product_purchase': 'https://wiki.seeedstudio.com/Grove-TemperatureAndHumidity_Sensor/',
-    'message': "Enter the Grove Pi+ GPIO pin to which the sensor is attached and select the sensor type.",
+    'url_manufacturer': [
+        'https://wiki.seeedstudio.com/Grove-Temperature_and_Humidity_Sensor_Pro/',
+        'https://wiki.seeedstudio.com/Grove-TemperatureAndHumidity_Sensor/',
+    ],
+    'message': "Enter the Grove Pi+ GPIO pin connected to the sensor and select the sensor type.",
 
     'options_enabled': [
         'gpio_location',
@@ -52,7 +52,11 @@ INPUT_INFORMATION = {
     'options_disabled': [
         'interface',
         'pre_output'
-        ],
+    ],
+
+    'dependencies_module': [
+        ('pip-pypi', 'grovepi', 'grovepi')
+    ],
 
     'custom_options': [
         {
@@ -60,8 +64,8 @@ INPUT_INFORMATION = {
             'type': 'select',
             'default_value': 0,
             'options_select': [
-                (0, 'Blue'),
-                (1, 'White')
+                (0, 'DHT11 (Blue)'),
+                (1, 'DHT22 (White)')
             ],
             'name': lazy_gettext('Sensor Type'),
             'phrase': 'Sensor type'
@@ -89,9 +93,6 @@ class InputModule(AbstractInput):
     """
     def __init__(self, input_dev, testing=False):
         """
-        :param gpio: gpio pin number
-        :type gpio: int
-
         Instantiate with the Pi and gpio to which the DHT11 output
         pin is connected.
 
@@ -99,6 +100,7 @@ class InputModule(AbstractInput):
         super(InputModule, self).__init__(input_dev, testing=testing, name=__name__)
 
         self.gpio = None
+        self.grovepi = None
         self.control = None
 
         self.temp_temperature = 0
@@ -116,7 +118,10 @@ class InputModule(AbstractInput):
         self.logger.info("DHT sensor initialization complete, pin={}, type={}.".format(self.gpio, self.sensor_type))
 
     def initialize_input(self):
+        import grovepi
         from mycodo.mycodo_client import DaemonControl
+
+        self.grovepi = grovepi
 
         self.gpio = int(self.input_dev.gpio_location)
 
@@ -132,10 +137,14 @@ class InputModule(AbstractInput):
         for _ in range(2):
             self.measure_sensor()
             if self.temp_dew_point is not None:
-                self.value_set(0, self.temp_temperature)
-                self.value_set(1, self.temp_humidity)
-                self.value_set(2, self.temp_dew_point)
-                self.value_set(3, self.temp_vpd)
+                if self.is_enabled(0):
+                    self.value_set(0, self.temp_temperature)
+                if self.is_enabled(1):
+                    self.value_set(1, self.temp_humidity)
+                if self.is_enabled(2):
+                    self.value_set(2, self.temp_dew_point)
+                if self.is_enabled(3):
+                    self.value_set(3, self.temp_vpd)
                 return self.return_dict  # success - no errors
             time.sleep(2)
 
@@ -153,17 +162,24 @@ class InputModule(AbstractInput):
                 self.setup()
             except Exception as except_msg:
                 self.logger.error(
-                    'Could not initialize sensor. Check if gpiod is running. Error: {msg}'.format(msg=except_msg))
-            import grovepi
+                    'Could not initialize sensor. Check if gpiod is running. '
+                    'Error: {msg}'.format(msg=except_msg))
+
             if isinstance(self.sensor_type, str):
                 self.sensor_type = int(self.sensor_type)
-            [self.temp_temperature, self.temp_humidity, retval] = grovepi.dht(self.gpio, self.sensor_type)
-            if retval == None:
-                self.logger.error("Error reading from DHT sensor: {}".format(retval))
-            self.logger.info("Temp: {}, Hum: {}".format(self.temp_temperature, self.temp_humidity))
+            [self.temp_temperature,
+             self.temp_humidity,
+             retval] = self.grovepi.dht(self.gpio, self.sensor_type)
+            if retval is None:
+                self.logger.error(
+                    "Error reading from DHT sensor: {}".format(retval))
+            self.logger.info("Temp: {}, Hum: {}".format(
+                self.temp_temperature, self.temp_humidity))
             if self.temp_humidity != 0:
-                self.temp_dew_point = calculate_dewpoint(self.temp_temperature, self.temp_humidity)
-                self.temp_vpd = calculate_vapor_pressure_deficit(self.temp_temperature, self.temp_humidity)
+                self.temp_dew_point = calculate_dewpoint(
+                    self.temp_temperature, self.temp_humidity)
+                self.temp_vpd = calculate_vapor_pressure_deficit(
+                    self.temp_temperature, self.temp_humidity)
         except Exception as e:
             self.logger.error("Exception raised when taking a reading: {err}".format(err=e))
         finally:
@@ -182,4 +198,3 @@ class InputModule(AbstractInput):
 
     def stop_input(self):
         return
-